+++ shaders/program/dh_water.glsl
@@ -127,7 +127,7 @@ void main() {
 		//	#else
 		//		foamAmount *= 0.7;
 		//	#endif
-		//	color.rgb = mix(color.rgb, vec3(0.75 + 0.25 * dayPercent), foamAmount * WATER_FOAM_AMOUNT * 2.5);
+		//	color.rgb = mix(color.rgb, vec3(0.75 + 0.25 * dayPercent), foamAmount * WATER_FOAM_AMOUNT * 2.0);
 		//#endif
 		
 		//// water needs to be more opaque in dark areas