+++ changelog.md
@@ -1,6 +1,7 @@
 - v1.4.0b
   - Replaced settings 'Shadows Enabled (Overworld/End)' with 'Shadows Type (Overworld/End)'
   - Disabled 'Fog Ignores Transparents' by default due to unexpected bugs
+  - Improved the color and shape of water caustics
 
 <br>